Moved powerup classes to a powerup package and made 2 pickup classes (coin and life). <layout
