+++ src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/IGameObjects.java
@@ -6,7 +6,7 @@
 package nl.joshuaslik.tudelft.SEM.control.gameObjects;
 
 import java.io.InputStream;
-import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.IModifier;
+import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.IModifier;
 import nl.joshuaslik.tudelft.SEM.control.viewController.viewObjects.ICircleViewObject;
 import nl.joshuaslik.tudelft.SEM.control.viewController.viewObjects.IImageViewObject;
 import nl.joshuaslik.tudelft.SEM.control.viewController.viewObjects.ILineViewObject;
