+++ src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/pickup/powerup/IDecorator.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup;
+package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup;
 
 import java.io.InputStream;
