+++ src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/pickup/powerup/bubble/EnumBubbleMods.java
@@ -1,3 +1,5 @@
+package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.bubble;
+
 ///*
 // * To change this license header, choose License Headers in Project Properties.
 // * To change this template file, choose Tools | Templates
