+++ src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/pickup/powerup/bubble/IBubbleModifier.java
@@ -3,9 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.bubble;
+package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.bubble;
 
-import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.IModifier;
+import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.IModifier;
 
 /**
  *
