+++ src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/pickup/powerup/player/AbstractPlayerModifierDecorator.java
@@ -3,10 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.player;
+package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.player;
 
-import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.IDecorator;
-import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.IModifier;
+import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.IDecorator;
+import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.IModifier;
 
 /**
  *
