+++ src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/pickup/powerup/player/EnumPlayerMods.java
@@ -1,3 +1,5 @@
+package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.player;
+
 ///*
 // * To change this license header, choose License Headers in Project Properties.
 // * To change this template file, choose Tools | Templates
